requirements.txt
# Core
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.10,<3.0
pydantic-settings>=2.7,<3.0

# LLM
anthropic>=0.49,<1.0

# Telegram
python-telegram-bot>=21,<22

# Web search
tavily-python>=0.5,<1.0

# Utilities
httpx>=0.28,<1.0
python-dotenv>=1.0,<2.0

# Logging & observability
structlog>=25,<26

# Testing
pytest>=8,<9
pytest-asyncio>=0.25,<1.0
